% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{align_proxies}
\alias{align_proxies}
\title{Align proxies}
\usage{
align_proxies(ld, e, vcf_ref = NULL, tempfile = NULL)
}
\arguments{
\item{ld}{output from get_ld_proxies}

\item{e}{output from extract_from_bcf}

\item{vcf_ref}{Optional, if provided then final dataset is aligned to reference}

\item{tempfile}{temporary files for use for extractions}
}
\value{
data frame
}
\description{
Takes output from get_ld_proxies and extract_from_bcf to make sure effect estimates are relative to the correct proxy alleles
}
