% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid_index.r
\name{create_ldref_sqlite}
\alias{create_ldref_sqlite}
\title{Create LD reference sqlite database for tags}
\usage{
create_ldref_sqlite(bfile, dbname, tag_r2 = 0.6)
}
\arguments{
\item{bfile}{path to plink file}

\item{dbname}{dbname to produce (overwrites existing if exists)}

\item{tag_r2}{minimum tag r2}
}
\description{
This is used for looking up proxies
}
