% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_index.r
\name{create_pval_index_from_vcf}
\alias{create_pval_index_from_vcf}
\title{Create pval index from GWAS-VCF file}
\usage{
create_pval_index_from_vcf(vcffile, maximum_pval, indexname)
}
\arguments{
\item{vcffile}{VCF filename}

\item{maximum_pval}{Maximum p-value to include. Default = 0.05}

\item{indexname}{index file name to create. Deletes existing file if exists.}
}
\description{
Create a separate file called \verb{<id>.pvali} which is used to speed up p-value queries.
}
