% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid_index.r
\name{create_rsidx_index}
\alias{create_rsidx_index}
\title{<brief desc>}
\usage{
create_rsidx_index(vcf, indexname)
}
\arguments{
\item{vcf}{<what param does>}

\item{indexname}{<what param does>}
}
\value{

}
\description{
<full description>
}
