% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid_index.r
\name{create_rsidx_sub_index}
\alias{create_rsidx_sub_index}
\title{Create new index from existing index using a subset of rsids}
\usage{
create_rsidx_sub_index(rsid, rsidx, newindex)
}
\arguments{
\item{rsid}{Vector of rsids}

\item{rsidx}{Existing index}

\item{newindex}{New index (Note: will delete existing file if exists)}
}
\value{
NULL, creates new index file
}
\description{
Note this requires a modified version of plink that allows ld-window-r2 flag for --r option.
Available here: https://github.com/explodecomputer/plink-ng
}
