% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.R
\name{create_vcf}
\alias{create_vcf}
\title{Create GWAS vcf}
\usage{
create_vcf(chrom, pos, nea, ea, snp = NULL, ea_af = NULL,
  effect = NULL, se = NULL, pval = NULL, n = NULL, ncase = NULL,
  name = NULL)
}
\arguments{
\item{dat}{data frame}

\item{snp_col}{snp_col}

\item{chrom_col}{chrom_col}

\item{pos_col}{pos_col}

\item{nea_col}{nea_col}

\item{ea_col}{ea_col}

\item{ea_af_col}{ea_af_col}

\item{effect_col}{effect_col}

\item{se_col}{se_col}

\item{pval_col}{pval_col}

\item{n_col}{n_col}

\item{ncase_col}{ncase_col}

\item{info_col}{info_col}

\item{z_col}{z_col}
}
\value{
vcf object
}
\description{
Create GWAS vcf
}
