% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Extract SNPs from bcf file}
\usage{
extract(bcf, snplist, tempname, proxies = "yes", bfile, vcf_ref,
  tag_kb = 5000, tag_nsnp = 5000, tag_r2 = 0.6)
}
\arguments{
\item{bcf}{vcf file name}

\item{snplist}{list of rs IDs or data frame of chrom and pos}

\item{tempname}{Temporary file}

\item{bfile}{ld reference panel}

\item{vcf_ref}{Optional, if provided then final dataset is aligned to reference}

\item{proxies="yes"}{If SNPs are absent then look for proxies (yes) or not (no). Can also mask all target SNPs and only return proxies (only), for testing purposes}

\item{tag_kb=5000}{Proxy parameter}

\item{tag_nsnp=5000}{Proxy parameter}

\item{tag_r2=0.6}{Proxy parameter}
}
\description{
Finds proxies if necessary
}
