% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_ld_proxies}
\alias{get_ld_proxies}
\title{Find LD proxies for a set of SNPs}
\usage{
get_ld_proxies(rsids, bcf, bfile, out, tag_kb = 5000, tag_nsnp = 5000,
  tag_r2 = 0.6)
}
\arguments{
\item{rsids}{list of rs IDs}

\item{bcf}{vcf file name}

\item{bfile}{ld reference panel}

\item{out}{temporary output file}

\item{tag_kb=5000}{Proxy parameter}

\item{tag_nsnp=5000}{Proxy parameter}

\item{tag_r2=0.6}{Proxy parameter}
}
\value{
data frame
}
\description{
Find LD proxies for a set of SNPs
}
