% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.R
\name{merge_vcf}
\alias{merge_vcf}
\title{Merge two GWAS VCF objects}
\usage{
merge_vcf(a, b)
}
\arguments{
\item{a}{VCF object}

\item{b}{VCF object}
}
\value{
SimpleList of VCF objects
}
\description{
Returns merged intersection of two VCF objects
}
