% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.r
\name{perform_coloc}
\alias{perform_coloc}
\title{Perform coloc}
\usage{
perform_coloc(vcf1, vcf2, chrompos)
}
\arguments{
\item{vcf1}{VCF object}

\item{vcf2}{VCF object}

\item{chrompos}{chrompos object (see input to parse_chrompos)}
}
\value{
Result from coloc
}
\description{
Perform coloc between two vcfs given a particular locus
}
