% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{proxy_match}
\alias{proxy_match}
\title{Extract SNPs from vcf file}
\usage{
proxy_match(vcf, rsid, bfile, proxies = "yes", tag_kb = 5000,
  tag_nsnp = 5000, tag_r2 = 0.6, threads = 1)
}
\arguments{
\item{vcf}{vcf file name}

\item{rsid}{list of rs IDs}

\item{bfile}{ld reference panel}

\item{proxies="yes"}{If SNPs are absent then look for proxies (yes) or not (no). Can also mask all target SNPs and only return proxies (only), for testing purposes}

\item{tag_kb=5000}{Proxy parameter}

\item{tag_nsnp=5000}{Proxy parameter}

\item{tag_r2=0.6}{Proxy parameter}

\item{threads=1}{}
}
\value{
data frame
}
\description{
Finds proxies if necessary
}
