% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{query_chrompos_vcf}
\alias{query_chrompos_vcf}
\title{Query chrompos from vcf object}
\usage{
query_chrompos_vcf(chrompos, vcf, id = NULL)
}
\arguments{
\item{chrompos}{Either vector of chromosome and position ranges e.g. "1:1000" or "1:1000-2000", or data frame with columns \code{chrom}, \code{start}, \code{end}.}

\item{vcf}{VCF object (e.g. from readVcf)}

\item{id}{If multiple GWAS datasets in the vcf file, the name (sample ID) from which to perform the filter}
}
\value{
VCF object
}
\description{
Query chrompos from vcf object
}
