% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_pval_bcftools}
\alias{query_pval_bcftools}
\title{Query p-value using bcftools}
\usage{
query_pval_bcftools(pval, vcffile, id = NULL)
}
\arguments{
\item{pval}{P-value threshold (NOT -log10)}

\item{vcffile}{Path to .vcf.gz GWAS summary data file}

\item{id}{If multiple GWAS datasets in the vcf file, the name (sample ID) from which to perform the filter}
}
\value{
vcf object
}
\description{
Query p-value using bcftools
}
