% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_pval_file}
\alias{query_pval_file}
\title{Query pval from vcf file}
\usage{
query_pval_file(pval, vcffile, id = NULL, build = "hg19")
}
\arguments{
\item{pval}{P-value threshold (NOT -log10)}

\item{vcffile}{Path to tabix indexed vcf file}

\item{id}{If multiple GWAS datasets in the vcf file, the name (sample ID) from which to perform the filter}

\item{build="hg19"}{}
}
\value{
VCF object
}
\description{
Query pval from vcf file
}
