% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_index.r, R/query.r
\name{query_pval_sqlite3}
\alias{query_pval_sqlite3}
\title{Query pval from file using pvali index}
\usage{
query_pval_sqlite3(pval, vcffile, id = NULL, pvali)

query_pval_sqlite3(pval, vcffile, id = NULL, pvali)
}
\arguments{
\item{pval}{pval threshold}

\item{vcffile}{Path to .vcf.gz GWAS summary data file}

\item{id}{If multiple GWAS datasets in the vcf file, the name (sample ID) from which to perform the filter}

\item{pvali}{Path to pval index file}
}
\value{
vcf object

vcf object
}
\description{
See create_pvali_index

See create_pvali_index
}
