% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{query_pval_vcf}
\alias{query_pval_vcf}
\title{Query based on p-value threshold from vcf}
\usage{
query_pval_vcf(pval, vcf, id = NULL)
}
\arguments{
\item{pval}{P-value threshold (NOT -log10)}

\item{vcf}{VCF object (e.g. from readVcf)}

\item{id}{If multiple GWAS datasets in the vcf file, the name (sample ID) from which to perform the filter}
}
\value{
VCF object
}
\description{
Query based on p-value threshold from vcf
}
