% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_index.r, R/query.r
\name{query_pvali}
\alias{query_pvali}
\title{Query pvali}
\usage{
query_pvali(pval, pvali)

query_pvali(pval, pvali)
}
\arguments{
\item{pval}{pval threshold}

\item{pvali}{Path to pval index file}
}
\value{
data frame

data frame
}
\description{
Query pvali

Query pvali
}
