% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{query_rsid_bcftools}
\alias{query_rsid_bcftools}
\title{Query}
\usage{
query_rsid_bcftools(rsid, vcffile, id = NULL)
}
\arguments{
\item{rsid}{Vector of rsids}

\item{vcffile}{Path to .vcf.gz GWAS summary data file}

\item{id}{If multiple GWAS datasets in the vcf file, the name (sample ID) from which to perform the filter}
}
\value{
VCF object
}
\description{
Query
}
