% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{query_rsid_rsidx}
\alias{query_rsid_rsidx}
\title{Query rsid from file using rsidx index}
\usage{
query_rsid_rsidx(rsid, vcffile, id = NULL, rsidx)
}
\arguments{
\item{rsid}{Vector of rsids}

\item{vcffile}{Path to .vcf.gz GWAS summary data file}

\item{id}{If multiple GWAS datasets in the vcf file, the name (sample ID) from which to perform the filter}

\item{rsidx}{Path to rsidx index file}
}
\value{
vcf object
}
\description{
See create_rsidx_index
}
