% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{query_rsid_vcf}
\alias{query_rsid_vcf}
\title{Query rsid from vcf object}
\usage{
query_rsid_vcf(rsid, vcf, id = NULL)
}
\arguments{
\item{rsid}{Vector of rsids}

\item{vcf}{VCF object (e.g. from readVcf)}

\item{id}{If multiple GWAS datasets in the vcf file, the name (sample ID) from which to perform the filter}
}
\value{
VCF object
}
\description{
Query rsid from vcf object
}
