% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{query_rsidx}
\alias{query_rsidx}
\title{Query rsidx}
\usage{
query_rsidx(rsid, rsidx)
}
\arguments{
\item{rsid}{Vector of rsids}

\item{rsidx}{Path to rsidx index file}
}
\value{
data frame
}
\description{
Query rsidx
}
