% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.R
\name{read_reference}
\alias{read_reference}
\title{Read in reference dataset}
\usage{
read_reference(reference_file, snplist, outfile, remove_dup_rsids = TRUE)
}
\arguments{
\item{reference_file}{<what param does>}

\item{snplist}{<what param does>}

\item{outfile}{<what param does>}

\item{remove_dup_rsids=TRUE}{<what param does>}
}
\value{
data frame
}
\description{
Read in reference dataset
}
