% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaries.r
\name{set_bcftools}
\alias{set_bcftools}
\title{Set bcftools binary location}
\usage{
set_bcftools(path = "")
}
\arguments{
\item{path}{If "" (default), then will use the MRCIEU/genetics.binaRies to get binaries that are appropriate for the detected operating system. Otherwise, provide the path to the bcftools binary. If NULL then will set the option to NULL.}
}
\value{
NULL, sets option 'tools_bcftools'
}
\description{
Set bcftools binary location
}
