% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaries.r
\name{set_plink}
\alias{set_plink}
\title{Set plink binary location}
\usage{
set_plink(path = "")
}
\arguments{
\item{path}{If "" (default), then will use the explodecomputer/genetics.binaRies to get binaries that are appropriate for the detected operating system. Otherwise, provide the path to the plink binary. If NULL then will set the option to NULL.}
}
\value{
NULL, sets option 'tools_plink'
}
\description{
Set plink binary location
}
