% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.r
\name{sqlite_ld_proxies}
\alias{sqlite_ld_proxies}
\title{Lookup LD proxies from sqlite database}
\usage{
sqlite_ld_proxies(rsids, dbfile, tag_r2)
}
\arguments{
\item{rsids}{List of rsids}

\item{dbfile}{path to dbfile}

\item{tag_r2}{minimum r2 value}
}
\value{
data frame
}
\description{
Lookup LD proxies from sqlite database
}
