% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{vcf_to_granges}
\alias{vcf_to_granges}
\title{Convert vcf format to granges format}
\usage{
vcf_to_granges(vcf, id = NULL)
}
\arguments{
\item{vcf}{Output from readVcf}

\item{id}{Only accepts one ID, so specify here if there are multiple GWAS datasets in the vcf}
}
\value{
GRanges object
}
\description{
Convert vcf format to granges format
}
