% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.r
\name{vcf_to_tibble}
\alias{vcf_to_tibble}
\title{Convert vcf format to tibble (data frame)}
\usage{
vcf_to_tibble(vcf, id = NULL)
}
\arguments{
\item{vcf}{Output from readVcf}

\item{id}{Only accepts one ID, so specify here if there are multiple GWAS datasets in the vcf}
}
\value{
GRanges object
}
\description{
Convert vcf format to tibble (data frame)
}
