% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{vcf_to_TwoSampleMR}
\alias{vcf_to_TwoSampleMR}
\title{Create exposure or outcome data format for TwoSampleMR from vcf}
\usage{
vcf_to_TwoSampleMR(vcf, type = "exposure")
}
\arguments{
\item{vcf}{VCF object}

\item{type}{="exposure" or "outcome"}
}
\value{
data frame
}
\description{
Create exposure or outcome data format for TwoSampleMR from vcf
}
