% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.r
\name{vcflist_overlaps}
\alias{vcflist_overlaps}
\title{Reduce list of VCFs to intersecting regions}
\usage{
vcflist_overlaps(vcflist, chrompos)
}
\arguments{
\item{vcflist}{List of VCF objects, or list of VCF filenames, or mix of VCF objects and filenames}

\item{chrompos}{Either vector of chromosome and position ranges e.g. "1:1000" or "1:1000-2000", or data frame with columns \code{chrom}, \code{start}, \code{end}.}
}
\value{
List of VCFs
}
\description{
Reduce list of VCFs to intersecting regions
}
