% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{write_out}
\alias{write_out}
\title{Write to neo4j csv format}
\usage{
write_out(x, basename, header = FALSE)
}
\arguments{
\item{x}{data frame}

\item{basename}{Filename to write to}

\item{header=FALSE}{Whether to also write a header}
}
\description{
Write to neo4j csv format
}
