% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsid_index.r
\name{create_rsidx_index_from_vcf}
\alias{create_rsidx_index_from_vcf}
\title{Create RSID index from VCF}
\usage{
create_rsidx_index_from_vcf(vcf, indexname)
}
\arguments{
\item{vcf}{VCF filename}

\item{indexname}{index file name to create. Deletes existing file if exists.}
}
\description{
Create RSID index from VCF
}
