% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.R
\name{create_vcf}
\alias{create_vcf}
\title{<brief desc>}
\usage{
create_vcf(chrom, pos, nea, ea, snp = NULL, ea_af = NULL,
  effect = NULL, se = NULL, pval = NULL, n = NULL, ncase = NULL,
  name = NULL)
}
\arguments{
\item{dat}{<what param does>}

\item{snp_col}{<what param does>}

\item{chrom_col}{<what param does>}

\item{pos_col}{<what param does>}

\item{nea_col}{<what param does>}

\item{ea_col}{<what param does>}

\item{ea_af_col}{<what param does>}

\item{effect_col}{<what param does>}

\item{se_col}{<what param does>}

\item{pval_col}{<what param does>}

\item{n_col}{<what param does>}

\item{ncase_col}{<what param does>}

\item{info_col}{<what param does>}

\item{z_col}{<what param does>}
}
\description{
<full description>
}
