% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.r
\name{create_vcf}
\alias{create_vcf}
\title{Create GWAS vcf}
\usage{
create_vcf(
  chrom,
  pos,
  nea,
  ea,
  snp = NULL,
  ea_af = NULL,
  effect = NULL,
  se = NULL,
  pval = NULL,
  n = NULL,
  ncase = NULL,
  name = NULL
)
}
\arguments{
\item{chrom}{chrom vector}

\item{pos}{pos vector}

\item{nea}{nea vector}

\item{ea}{ea vector}

\item{snp}{Optional vector}

\item{ea_af}{Optional vector}

\item{effect}{Optional vector}

\item{se}{Optional vector}

\item{pval}{Optional vector}

\item{n}{Optional vector}

\item{ncase}{Optional vector}

\item{name}{Optional vector}
}
\value{
vcf object
}
\description{
Create GWAS vcf
}
