% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_from_bcf}
\alias{extract_from_bcf}
\title{Extract SNPs from vcf}
\usage{
extract_from_bcf(snplist, bcf, out,
  sel = "\%CHROM \%POS \%ID \%REF \%ALT \%B \%SE \%PVAL \%N \%AF\\n")
}
\arguments{
\item{snplist}{list of rs IDs or table of chrom and pos}

\item{bcf}{bcf file name}

\item{out}{temporary filename}

\item{sel='%CHROM}{%POS %ID %REF %ALT %B %SE %PVAL %N %AF\n' What to extract from bcf}
}
\value{
data frame
}
\description{
Extract SNPs from vcf
}
