% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_from_bcf}
\alias{extract_from_bcf}
\title{Extract SNPs from vcf}
\usage{
extract_from_bcf(snplist, bcf, out,
  sel = "\%CHROM \%POS \%ID \%REF \%ALT \%EFFECT \%SE \%L10PVAL \%N \%AF\\n",
  logpval = FALSE)
}
\arguments{
\item{snplist}{list of rs IDs or table of chrom and pos}

\item{bcf}{bcf file name}

\item{out}{temporary filename}

\item{logpval}{Whether to return -log10 or standard p-value. Default is standard (though it is stored as logged in bcf to retain precision)}

\item{sel='%CHROM}{%POS %ID %REF %ALT %EFFECT %SE %L10PVAL %N %AF\n' What to extract from bcf}
}
\value{
data frame
}
\description{
Extract SNPs from vcf
}
