% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasglue.R
\name{gwasvcf_to_summaryset}
\alias{gwasvcf_to_summaryset}
\title{Create a SummarySet}
\usage{
gwasvcf_to_summaryset(vcf)
}
\arguments{
\item{vcf}{Path or URL to GWAS-VCF file or VCF object e.g. output from VariantAnnotation::readVcf, gwasvcftools::query_vcf or [query_gwas()]}
}
\description{
Returns a gwasglue2 SummarySet object
}
