% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{parse_chrompos}
\alias{parse_chrompos}
\title{Parse chromosome:position}
\usage{
parse_chrompos(chrompos, radius = NULL)
}
\arguments{
\item{chrompos}{Either vector of chromosome and position ranges e.g. "1:1000" or "1:1000-2000", or data frame with columns \code{chrom}, \code{start}, \code{end}.}

\item{radius}{Add radius to the specified positions. Default = NULL}
}
\value{
GRanges object
}
\description{
Takes data frame or vector of chromosome position ranges and parses to granges object
}
