% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{query_chrompos_file}
\alias{query_chrompos_file}
\title{Query vcf file, extracting by chromosome and position}
\usage{
query_chrompos_file(chrompos, vcffile, id = NULL, build = "GRCh37")
}
\arguments{
\item{chrompos}{Either vector of chromosome and position ranges e.g. "1:1000" or "1:1000-2000", or data frame with columns \code{chrom}, \code{start}, \code{end}.}

\item{vcffile}{Path to .vcf.gz GWAS summary data file}

\item{id}{If multiple GWAS datasets in the vcf file, the name (sample ID) from which to perform the filter}

\item{build}{Default="GRCh37" Build of vcffile}
}
\value{
VCF object
}
\description{
Query vcf file, extracting by chromosome and position
}
