% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_gwas}
\alias{query_gwas}
\title{Query data from vcf file}
\usage{
query_gwas(vcf = NULL, vcffile = NULL, chrompos = NULL,
  rsid = NULL, pval = NULL, id = NULL, build = "hg19")
}
\arguments{
\item{vcf}{VCF object e.g. output from VariantAnnotation::readVcf or gwasvcftools::query_vcf}

\item{vcffile}{Path to .vcf.gz GWAS summary data file}

\item{chrompos}{Either vector of chromosome and position ranges e.g. "1:1000" or "1:1000-2000", or data frame with columns `chrom`, `start`, `end`.}

\item{rsid}{Vector of rsids}

\item{pval}{P-value threshold (NOT -log10)}

\item{id}{If multiple GWAS datasets in the vcf file, the name (sample ID) from which to perform the filter}

\item{build="hg19"}{Build of vcffile}
}
\value{
vcf object
}
\description{
Read in GWAS summary data with filters on datasets (if multiple datasets per file) and/or chromosome/position, rsids or pvalues. Chooses most optimal choice for the detected operating system. Typically chrompos searches are the fastest. On Windows, rsid or pvalue filters from a file will be slow.
}
