% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{query_gwas}
\alias{query_gwas}
\title{Query data from vcf file}
\usage{
query_gwas(
  vcf,
  chrompos = NULL,
  rsid = NULL,
  pval = NULL,
  id = NULL,
  rsidx = NULL,
  build = "GRCh37",
  os = Sys.info()[["sysname"]],
  proxies = "no",
  bfile = NULL,
  dbfile = NULL,
  tag_kb = 5000,
  tag_nsnp = 5000,
  tag_r2 = 0.6,
  threads = 1
)
}
\arguments{
\item{vcf}{VCF object e.g. output from VariantAnnotation::readVcf or gwasvcftools::query_vcf}

\item{chrompos}{Either vector of chromosome and position ranges e.g. "1:1000" or "1:1000-2000", or data frame with columns `chrom`, `start`, `end`.}

\item{rsid}{Vector of rsids}

\item{pval}{P-value threshold (NOT -log10)}

\item{id}{If multiple GWAS datasets in the vcf file, the name (sample ID) from which to perform the filter}

\item{rsidx}{Path to rsidx index file}

\item{build}{="GRCh37" Build of vcffile}

\item{os}{The operating system. Default is as detected. Determines the method used to perform query}

\item{proxies}{="no" If SNPs are absent then look for proxies (yes) or not (no). Can also mask all target SNPs and only return proxies (only), for testing purposes. Currently only possible if querying rsid.}

\item{bfile}{=path to plink bed/bim/fam ld reference panel}

\item{dbfile}{=path to sqlite tag snp database}

\item{tag_kb}{=5000 Proxy parameter}

\item{tag_nsnp}{=5000 Proxy parameter}

\item{tag_r2}{=0.6 Proxy parameter}

\item{threads}{=1 NUmber of threads}
}
\value{
vcf object
}
\description{
Read in GWAS summary data with filters on datasets (if multiple datasets per file) and/or chromosome/position, rsids or pvalues. Chooses most optimal choice for the detected operating system. Typically chrompos searches are the fastest. On Windows, rsid or pvalue filters from a file will be slow.
}
