% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R, R/harmonise.R
\name{write_out}
\alias{write_out}
\title{Write to neo4j csv format}
\usage{
write_out(harmonised, path)

write_out(harmonised, path)
}
\arguments{
\item{harmonised}{Output from /code{harmonise_against_ref}}

\item{path}{Path to write out json file and txt file}

\item{x}{data frame}

\item{basename}{Filename to write to}

\item{header=FALSE}{Whether to also write a header}
}
\description{
Takes raw files and aligns them to reference. Important if files don't have chr:pos already
}
